Fixed constructor, readded getter

+++ kyoto/src/main/java/uk/ac/ic/kyoto/annex1reduce/EUCountry.java
@@ -13,7 +13,7 @@ import uk.ac.imperial.presage2.core.messaging.Input;
 public class EUCountry extends AbstractCountry {
 	
 	public EUCountry(UUID id, String name,String ISO, double landArea, double arableLandArea, double GDP,
-			double GDPRate, float availiableToSpend, long emissionsTarget, long carbonOffset,
+			double GDPRate, long availiableToSpend, long emissionsTarget, long carbonOffset,
 			long energyOutput, long carbonOutput) {
 		
 		super(id, name, ISO, landArea, arableLandArea, GDP,
